requirements.txt
Flask
Flask-Cors
pymongo
python-dotenv
scikit-learn
pandas
numpy
fuzzywuzzy
python-Levenshtein
gunicorn